refactor: realocando getProducts para routers
