+++ src/routers/products.routers.js
@@ -1,0 +1,9 @@
+import { Router } from "express";
+
+import { getProducts } from "../controllers/products.controllers.js";
+
+const router = Router();
+
+router.get("/products", getProducts );
+
+export default router;
