+++ src/routers/users.routers.js
@@ -1,6 +1,6 @@
 import { Router } from "express";
 
-import { postSingUp, postSingIn, deleteGoOut, getProducts } from "../controllers/users.controller.js"
+import { postSingUp, postSingIn, deleteGoOut } from "../controllers/users.controller.js"
 import { singUpMD } from "../middleware/singUp.middleware.js";
 import { singInMD } from "../middleware/singIn.middleware.js";
 
@@ -9,6 +9,5 @@ const router = Router();
 router.post("/sing-up", singUpMD, postSingUp);
 router.post("/sing-in", singInMD, postSingIn);
 router.delete( "/go-out", deleteGoOut );
-router.get("/products", getProducts )
 
 export default router;
